fix: remove destroy comment
Remove comment as this destructor is really here to appease the compiler
and make it clear that we don't want to destroy anything explicitly.
The underlying storage space for the string may now be in shared memory
and no longer under control of just this process. When deletion is
supported, destruction should be implemented.