Use node-pre-gyp to publish pre-compiled packages.
These pre-compiled packages are sensitive to Linux version, Node
version, and system libraries (GLIBC in particular). There is a
procedure at the bottom of the README that details the release
process. This must be done on an architecture that's going to match the
target architecture this is going to be used on.  In particular, the
libc/GLIBC versions must match or the module will fail with a:

    Error: /lib64/libc.so.6: version `GLIBC_2.17' not found

or similar message.

This commit also:

- Adds a test file and regression test for the old format.  There's
something not entirely great about it so it's marked as "skipped" for now.

- Uses node-pre-gyp-github to upload binary releases to Github instead of
S3.

- Adds emacs backup files to gitignore.