Touchups of Dmitry Savransky's excellent work.
Removed `keys()` method in favor of the more standard `Object.keys()`
construction which Dmitry's work enabled.

Continue to disallow undefined and null values. These are valid values
for normal Javascript objects:

    > obj = {key1: undefined, key2: null}
    { key1: undefined, key2: null }

That structure won't be preserved in the current system so the writer
and reader can have different views of the object. Deletion works now,
so setting to null should not be necessary.

Send back what should be the proper query properties for various items
in the map. Remove delegation to v8, as it shouldn't be able to find
anything more.

Convert some throws on deletions to return values. This mirrors real
objects. Retail throw on symbols as this is necessarily different from
real objects.

Create tests to cover all this stuff.

Refactors: consolidate method membership.