Skip compatibility tests on OSX
Boost's memory mapped file routines are not compatible between OSX and
Linux.

This adds a Linux-build binary for the compatibility test and skips
running the test on OSX.  Travis runs these tests under Linux, so keep
the test for that platform.  No clue if this is going to work under
Windows or not.