Support numeric props, change deletion semantics
Breaking API change: when deleting a property on a read-only object,
an error is now thrown. Before, this would fail silently.

This commit also adds support for getting/setting properties that are
just numbers. Seems like v8 assumes any number (even strings made out
of digits) is an index access. This change translates those to object
accesses. Use of mmap-object as an array isn't supported otherwise.

Also deal with cases when inspect and/or symbol properties are
accessed.