Allow close() to be called asynchronously
For particularly large files (5GB+) that end up loading a lot of pages
into memory (i.e. many reads that end up pulling a lot of data off the
backing store), msync/munmap may take a bit of time. One strace of
system calls showed 250ms for this size of file and there are
anecdotes of a second or longer.

When close() is called with a callback, the close operation will occur
in a separate thread. This should allow the main thread to continue
with operations while the msync/munmap takes its time.