Address Call() deprecation, move error handling into CloseWorker
Compiling with Node 9.6.0 gives this message:

  ../mmap-object.cc:552:17: warning: 'Call' is deprecated [-Wdeprecated-declarations]
        callback->Call(1, argv);
                  ^
  ../node_modules/nan/nan.h:1567:3: note: 'Call' has been explicitly marked deprecated here
    NAN_DEPRECATED inline v8::Local<v8::Value>
    ^
  ../node_modules/nan/nan.h:98:40: note: expanded from macro 'NAN_DEPRECATED'
  # define NAN_DEPRECATED __attribute__((deprecated))

It's deprecated because that Call() doesn't happen in a proper async
environment, so no ticks occur and no events are emitted. The error
handling is now moved inside the CloseWorker for the AsyncWorker to
handle. This means the callback is now properly async in error cases
as well as in non-error cases.

It also simplifies the code somewhat as the caller no longer creates a
new callback. Instead, AsyncWorker handles that transparently.

The non-async case is made more generic. Instead of checking for a
closed file, it checks to see if the CloseWorker registered any
error. If so, it'll throw that error.

Setting the map_seg to NULL and closed to true is moved inside the
worker. This gets those operations closer to where they actually
happen. Since the CloseWorker is initialized with the map and not the
map_seg/closed, this has to happen in the CloseWorker in order to
maintain live references.