Persist the map object in the async closeworker
There's a chance that the GC might run between the time the
CloseWorker is instantiated and the point at which its Execute()
method is called. If the SharedMap object goes out of scope in JS
immediately after the close() call is made, the map object may be
trashed and the call to Execute() results in a crash.

Making map persistent in the CloseWorker ensures that it'll stick
around as long as the worker is around. This will prevent the crash.

Version is bumped to cover. All packages are updated, a test is added.