Performance enhancements
A lot of time was being spent checking if an accessed property was a
method. Replacing the set<string> lookup with Aho Corasick halved the
latency of the critical path.

Moving the part that checks for the inspect method into the isMethod()
clause added a little bit of speedup as well.

Added propertyIsEnumerable as a method, because it is.

Version is bumped to 1.3.8