Support buffers as values
The v8 interface to Buffers allows for a memory pointer to be passed
back to Javascript without being copied (only a new buffer structure
needs to be allocated). This differs from strings which have to be
copied as they are handled entirely by v8's memory management system.

Rough benchmarks on reads show that for 2k-byte values, strings and
buffers perform about equally. For 20k-byte values, buffers are 300%
faster. For 200k-byte values, buffers are 2000% faster.

The memory handed to NewBuffer is deleted on GC by default. NullFreer
avoids the delete and the resulting crash (shared memory can't be
deleted).

Turns out setting a cell value can be simplified as passing a
shared_string to a copy constructor does the right thing in both
cases. Only the length needs to be set and there's no real loss in
setting it for a string (when it's always 0).

Cell::operator string() wasn't being used for anything.  Remove it.