Keep track of file version
This introduces the notion of file versions. All files created from
here on out are considered version 1. Previous formats are considered
version 0. Currently, version 1 and version 0 are considered
compatible.

Very old "version 0" files may not compatible with current or future
releases. All version 1 files are compatible with the current release.

Version checking is only done on file open so this logic should not
have any impact on the critical path.