Drop use of v8::External()
The usage of v8::External() associates the SharedMap object with a new
JS variable which is itself associated with the call handler on the
iter_template. After the template is used that variable is a candidate
for garbage collection. This means the SharedMap may be reclaimed and
destroyed. When that occurs, the iterator state runs off the rails and
a crash occurs.

Passing the wrapped object instead works fine and avoids this whole
situation.