Refactors, macOS updates
The README now reflects the actual base address used.

Handling shared memory on a Mac seems like it's going to be relatively
complex. In the future, a super-master lock can be kept at a known
base address (and at some harmonic locations) with a magic value as
identifier. Once found, the master location can be locked globally and
used to index mutexes for specific loaded files (by filename
perhaps).

So any open file would first consult the master location for a pointer
to the specific location in memory where the mutexes lay. Then that
location would be used for the shared mutexes for that file.

But none of this exists currently.

This commit does have one important fix in Open(): deallocating the
mutexes at another point that was overlooked.