Support Arm architecture
Also upgrade dependencies to latest versions. And do a little
whitespace elision.

With the glibc relabelings, Arm builds would fail:

  /usr/bin/ld: Release/obj.target/mmap-object.node: No symbol version section for versioned symbol `memcpy@GLIBC_2.2.5'
  /usr/bin/ld: final link failed: Nonrepresentable section on output

This addresses that situation.

Also, building with GCC 8 doesn't work so well. GCC 6 had no issues.

The format of the resulting file is dependent on the architecture so
allow for that.