Update packages, remove failing test
Combined `ncu -u` and `npm audit fix` to get a new package list.

A couple errors popped up when running on MacOS 11.2.1 (with node
v14.12.0):

  1) mmap-object
       Still can read old format
         "before all" hook for "reads string properties":
     Error: File /Users/allen/src/mmap-object/testdata/previous-format-darwin-x64.bin appears to be corrupt (1).
      at Context.<anonymous> (test/test-mmap-object.js:484:24)
      at processImmediate (internal/timers.js:461:21)

  2) mmap-object
       Still can read old format
         "after all" hook for "reads number properties":
     TypeError: Cannot read property 'close' of undefined
      at Context.<anonymous> (test/test-mmap-object.js:488:22)
      at processImmediate (internal/timers.js:461:21)

I'm guessing this might have something to do with the OS
version. Since the old version of the file format is pushing 5 years
now, I'm removing the test.